configure_vm_ssh.sh: don't overwrite /vagrant/vm/ssh/config if it exists

+++ setup/steps/configure_vm_ssh.sh
@@ -7,7 +7,8 @@ set -euo pipefail
 #
 if [[ $MAGAOX_ROLE == vm ]]; then
   mkdir -p /vagrant/vm/ssh
-  cat <<'HERE' | sudo tee /vagrant/vm/ssh/config
+  if [[ ! -e /vagrant/vm/ssh/config ]]; then
+    cat <<'HERE' | sudo tee /vagrant/vm/ssh/config
 IdentityFile /vagrant/vm/ssh/magaox_ssh_key
 Host aoc
   HostName exao1.magao-x.org
@@ -20,6 +21,7 @@ Host icc
 Host *
   User YOURUSERNAME
 HERE
+  fi
   mkdir -p /home/vagrant/.ssh
   if [[ ! -e /home/vagrant/.ssh/known_hosts ]]; then
       cat <<'HERE' | sudo tee /home/vagrant/.ssh/known_hosts
